refactor(library/Attempt): aligns file structure with level 2 namespace in "Attempt/Fresh"
