+++ src/library/Attempt/Fresh/synchronous.ts
